fix(ui): surface the Federation ID at top level, ungate live federation tiles, drop stale Awaiting-Edge chrome
On-device review of the 2.9.6 build (S21U) surfaced four issues:

- The Federation ID card (persist identity aggregate, this release) was
  buried in the hub → Identity sub-screen — behind a tile that still
  wore a stale "Coming soon" badge. The card is now a shared component
  (ui/components/FederationIdCard.kt) rendered at TOP LEVEL on both the
  Network overview (NetworkOpsScreen) and the federation hub
  (NetworkScreen), directly under the signer-key card. NetworkViewModel
  gains the aggregate fetch (null on graceful 503).
- All ten federation tiles wore unconditional "Coming soon" badges even
  though every sub-screen has been live since T-E/T-E-D — badges
  removed; tiles navigate directly. (Whole-screen placeholders for
  genuinely-unbuilt surfaces — Delegation/Participate/AgentsList —
  untouched.)
- "Awaiting Edge 1.0" chrome was stale (Edge 2.1.1 shipped): removed
  the QR/ratchet placeholder row from the hub identity card and the
  per-cell caption in the stats strip (values stay an honest "—" until
  real edge metrics wire in).
- NetworkOpsScreen: Scaffold TopAppBar title was clipped by the app
  shell's floating logo ("Network" → "ork") — replaced with the hub's
  in-content title pattern + Surface (the removed Scaffold was painting
  the theme background; its absence was the white-background
  regression) + logo clearance; the Data-directory row no longer
  collides label/value (stacked label + wrapping mono path).

Verified on-device: aggregate key, Ed25519 short form, and
Signing/PQC/Content-encryption chips render on both surfaces; dark
background restored; tiles navigate.
